Refactor move myinfo to admin
